tests/test_livres.py
import unittest
from app import create_app, db
from app.models import Livre


class TestLivres(unittest.TestCase):

    def setUp(self):
        self.app = create_app()
        self.client = self.app.test_client()
        self.app.app_context().push()
        db.create_all()

    def tearDown(self):
        db.session.remove()
        db.drop_all()


class TestLivres(unittest.TestCase):

    def setUp(self):
        self.app = create_app()
        self.client = self.app.test_client()
        self.app.app_context().push()
        db.create_all()

    def tearDown(self):
        db.session.remove()
        db.drop_all()

    def test_ajouter_livre(self):
        response = self.client.post(
            "/livres", json={"titre": "Livre Test", "auteur": "Auteur Test"}
        )
        self.assertEqual(response.status_code, 308)

    """  self.assertIn("Livre ajouté avec succès", response.get_json()["message"])
 """

    def test_lister_livres(self):
        self.client.post("/livres/", json={"titre": "Livre 1", "auteur": "Auteur 1"})

        response = self.client.get("/livres/")
        self.assertEqual(response.status_code, 200)
        """ self.assertIn(b"Livre1", response.data) """


""" if __name__ == "__main__":
    unittest.main()
 """
